addOrderTableAndContactView

+++ MedicalTools/MedicalTools/Controllers/CustomerController.cs
@@ -9,30 +9,19 @@ namespace MedicalTools.Controllers
 {
     public class CustomerController : Controller
     {
-        
+
         private readonly ApplicationContext _db;
         public CustomerController(ApplicationContext db)
         {
             _db = db;
         }
-        public class cartProductTuples
-        {
-            public Cart Cart { get; set; }
-            public Product Product { get; set; }
-        }
-		public class cartProductOrderTuples
-		{
-			public Cart Cart { get; set; }
-			public Product Product { get; set; }
-            public Order Order { get; set; }
-        }
-		public IActionResult Index()
+        public IActionResult Index()
         {
             return View();
         }
 
 
-        public IActionResult Products( int id )
+        public IActionResult Products(int id)
         {
             return View();
         }
@@ -45,32 +34,26 @@ namespace MedicalTools.Controllers
         {
 
             // id for current user must fill from session
-            string? userJson = HttpContext.Session.GetString("LiveUSer");
+            string? userJson = HttpContext.Session.GetString("LiveUser");
             var user = JsonConvert.DeserializeObject<User>(userJson);
-            if (true)
+            if (user != null)
             {
-				var products = _db.cart.Include(p=>p.product)
-				.Where(u => u.UserId == user.ID).ToList();
+                var products = _db.cart.Include(p => p.product)
+                .Where(u => u.UserId == user.ID).ToList();
 
-				return View(products);
-			}
+                return View(products);
+            }
 
-			//var cartsWithId = _db.cart
-            //.Where(u => u.UserId == user.ID).ToList();
+            TempData["ReturnUrl"] = Url.Action("Cart", "Customer");
+            return RedirectToAction("Index", "Login");
 
-
-			return RedirectToAction("Index","Login");
         }
         public IActionResult Checkout()
         {
-            string? userJson = HttpContext.Session.GetString("LiveUSer");
+            string? userJson = HttpContext.Session.GetString("LiveUser");
             var user = JsonConvert.DeserializeObject<User>(userJson);
             var orders = _db.cart.Include(p => p.product)
                 .Where(u => u.UserId == user.ID).ToList();
-            if (true)
-            {
-                return View(orders);
-            }
 
             Order order = new Order();
             double totalPrice = 0;
@@ -84,48 +67,70 @@ namespace MedicalTools.Controllers
             order.Name = user.Name;
             order.Location = user.location;
             order.Email = user.Email;
-
-            return View(order);
+            order.user = user;
+            order.user.carts = _db.cart.Where(u => u.UserId == user.ID).ToList();
+            if (user != null)
+            {
+                return View(order);
+            }
+            TempData["ReturnUrl"] = Url.Action("Checkout", "Customer");
+            return RedirectToAction("Index", "Login");
         }
         [HttpPost]
         public IActionResult Checkout(Order ord)
         {
-            var obj = _db.payments;
+            var obj = _db.payments.ToList();
+            string? userJson = HttpContext.Session.GetString("LiveUser");
+            var user = JsonConvert.DeserializeObject<User>(userJson);
+
+            //ord.Total = 0;
+            //var cart = _db.cart
+            //    .Where(u=>u.UserId == user.ID)
+            //    .ToList();
+            //foreach(var cartItem in cart)
+            //{
+            //    ord.Total += cart
+            //}
+            //ord.Total = _db.cart.Where(u=> u.UserId == user.)
+            ord.userId = user.ID;
             foreach (var item in obj)
             {
-                if (item.cardNo == ord.card)
-
+                if (item.cardNo == ord.card && item.Password == ord.password)
                 {
-                        if (item.Password == ord.password)
-                    {
-                        _db.orders.Add(ord);
-                        _db.SaveChanges();
-                        int id = 1; // id for current user must fill from session 
-                        var cartsToDelete = _db.cart.Where(cart => cart.UserId == id);
+                    _db.orders.Add(ord);
+                    _db.SaveChanges();
 
-                        _db.cart.RemoveRange(cartsToDelete);
-                        _db.SaveChanges();
+                    var cartsToDelete = _db.cart.Where(cart => cart.UserId == user.ID);
+                    _db.cart.RemoveRange(cartsToDelete);
+                    _db.SaveChanges();
 
-                        TempData["success"] = "Payment processed successfully";
-                        return RedirectToAction("Index");
-                    }
-                    else
-                    {
-                        TempData["error"] = "Payment does't work";
-                        return View();
-                    }
-                }
-                else
-                {
-                    TempData["error"] = "Payment does't work";
-                    return View();
+                    TempData["success"] = "Payment processed successfully";
+                    return RedirectToAction("Index");
                 }
             }
             TempData["error"] = "Payment does't work";
             return View();
         }
 
-		public IActionResult Login()
+        public IActionResult Order()
+        {
+            string? userJson = HttpContext.Session.GetString("LiveUSer");
+            var user = JsonConvert.DeserializeObject<User>(userJson);
+            if (user != null)
+            {
+                var orders = _db.orders.Include(u => u.user)
+                    .Where(o => o.userId == user.ID).ToList();
+
+                return View(orders);
+            }
+            TempData["ReturnUrl"] = Url.Action("Order", "Customer");
+            return RedirectToAction("Index", "Login");
+        }
+        public IActionResult Contact()
+        {
+            return View();
+        }
+        public IActionResult Login()
         {
             return View();
         }
@@ -134,11 +139,6 @@ namespace MedicalTools.Controllers
             return View();
         }
         public IActionResult About()
-        {
-            return View();
-        }
-
-        public IActionResult Contact()
         {
             return View();
         }
